- added `defaultFilter.vert` - moved all filters to the filter folder rather than having them separated - added baseTexture deprecation - renamed `myTexture` value to `uSampler` - added a simple migration function - added wrapMode deprecation
