+++ src/filters/color-matrix/colorMatrixFilter.frag
@@ -7,7 +7,7 @@ out vec4 fragColor;
 uniform float uColorMatrix[20];
 uniform float uAlpha;
 
-uniform sampler2D myTexture;
+uniform sampler2D uSampler;
 
 float rand(vec2 co)
 {
@@ -16,7 +16,7 @@ float rand(vec2 co)
 
 void main()
 {
-    vec4 color = texture(myTexture, vTextureCoord);
+    vec4 color = texture(uSampler, vTextureCoord);
     float randomValue = rand(gl_FragCoord.xy * 0.2);
     float diff = (randomValue - 0.5) *  0.5;
