+++ src/filters/color-matrix/colorMatrixFilter.wgsl
@@ -21,7 +21,7 @@ struct ColorMatrixUniforms {
 @group(0) @binding(0) var<uniform> globalUniforms : GlobalUniforms;
 
 @group(1) @binding(0) var<uniform> gfu: GlobalFilterUniforms;
-@group(1) @binding(1) var myTexture: texture_2d<f32>;
+@group(1) @binding(1) var uSampler: texture_2d<f32>;
 @group(1) @binding(2) var mySampler : sampler;
 @group(1) @binding(3) var backTexture: texture_2d<f32>;
 @group(2) @binding(0) var<uniform> colorMatrixUniforms : ColorMatrixUniforms;
@@ -61,7 +61,7 @@ fn mainFragment(
 ) -> @location(0) vec4<f32> {
 
 
-  var c = textureSample(myTexture, mySampler, uv);
+  var c = textureSample(uSampler, mySampler, uv);
   
   if (colorMatrixUniforms.uAlpha == 0.0) {
     return c;
