+++ src/filters/displacement/displacement.frag
@@ -4,7 +4,7 @@ in vec2 vFilterUv;
 
 out vec4 fragColor;
 
-uniform sampler2D myTexture;
+uniform sampler2D uSampler;
 uniform sampler2D mapTexture;
 
 uniform vec4 filterArea;
@@ -21,5 +21,5 @@ vec4 map = texture(mapTexture, vFilterUv);
     
     vec2 offset = inputSize.zw * (rotation * (map.xy - 0.5)) * scale; 
 
-    fragColor = texture(myTexture, clamp(vTextureCoord + offset, inputClamp.xy, inputClamp.zw));
+    fragColor = texture(uSampler, clamp(vTextureCoord + offset, inputClamp.xy, inputClamp.zw));
 }
