--- src/rendering/filters/noise/noise.vert
@@ -1,28 +1,0 @@
-in vec2 aPosition;
-out vec2 vTextureCoord;
-
-uniform globalUniforms {
-  mat3 projectionMatrix;
-  mat3 worldTransformMatrix;
-  float worldAlpha;
-};
-
-uniform vec4 inputSize;
-uniform vec4 outputFrame;
-
-vec4 filterVertexPosition( void )
-{
-    vec2 position = aPosition * outputFrame.zw + outputFrame.xy;
-    return vec4((projectionMatrix * vec3(position, 1.0)).xy, 0.0, 1.0);
-}
-
-vec2 filterTextureCoord( void )
-{
-    return aPosition * (outputFrame.zw * inputSize.zw);
-}
-
-void main(void)
-{
-    gl_Position = filterVertexPosition();
-    vTextureCoord = filterTextureCoord();
-}
