+++ src/rendering/filters/shared/Filter.ts
@@ -1,11 +1,11 @@
-import { Shader } from '../renderers/shared/shader/Shader';
-import { State } from '../renderers/shared/state/State';
+import { Shader } from '../../renderers/shared/shader/Shader';
+import { State } from '../../renderers/shared/state/State';
 
-import type { RenderSurface } from '../renderers/gpu/renderTarget/GpuRenderTargetSystem';
-import type { ShaderWithResourcesDescriptor } from '../renderers/shared/shader/Shader';
-import type { BLEND_MODES } from '../renderers/shared/state/const';
-import type { Texture } from '../renderers/shared/texture/Texture';
-import type { FilterSystem } from './shared/FilterSystem';
+import type { RenderSurface } from '../../renderers/gpu/renderTarget/GpuRenderTargetSystem';
+import type { ShaderWithResourcesDescriptor } from '../../renderers/shared/shader/Shader';
+import type { BLEND_MODES } from '../../renderers/shared/state/const';
+import type { Texture } from '../../renderers/shared/texture/Texture';
+import type { FilterSystem } from './FilterSystem';
 
 export interface FilterOptions extends ShaderWithResourcesDescriptor
 {
@@ -94,7 +94,7 @@ export class Filter extends Shader
         this.blendRequired = options.blendRequired;
 
         this.addResource('globalUniforms', 0, 0);
-        this.addResource('myTexture', 1, 1);
+        this.addResource('uSampler', 1, 1);
     }
 
     /**
