+++ src/rendering/renderers/gl/shader/migrateFragmentFromV7toV8.ts
@@ -1,0 +1,14 @@
+export function migrateFragmentFromV7toV8(fragmentShader: string): string
+{
+    fragmentShader = fragmentShader
+        .replaceAll('texture2D', 'texture')
+        .replaceAll('gl_FragColor', 'fragColor')
+        .replaceAll('varying', 'in');
+
+    fragmentShader = `
+        out vec4 fragColor;
+    ${fragmentShader}
+    `;
+
+    return fragmentShader;
+}
